Add AuditLogs resource (#17)
Add `AuditLogs` resource on the client as the `auditlogs` property. It does not support create/update/delete, so it will raise an error when any of those methods are called.

Also changed the version logic in `setup.py` to use the long version, which will add alpha/beta/RC versions to the end of the package when building so that the distribution to test-pypi can be tested more than once per version; just change the `serial` part and a new version will be generated, so the test-pypi build and upload script will work each time.

This is `v1.2.0-rc.1`